+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Restored usage of ENABLE_VR to fix compilation errors on some platforms.
 - Only call SetDirty on an object when actually modifying it in SRP updater utility
 - Set depthSlice to -1 by default on SetRenderTarget() to clear all slices of Texture2DArray by default.
+- Replace usage of PlayerSettings.virtualRealitySupported for 2020.1 by non-deprecated code.
 
 ### Fixed
 - Fixed compile errors for platforms with no VR support